refactor: Nest folders tools.
