+++ src/tools/folders/index.ts
@@ -1,0 +1,5 @@
+import "./create-folder";
+import "./delete-folder";
+import "./get-folder";
+import "./list-folders";
+import "./update-folder";
