+++ src/tools/folders/list-folders.ts
@@ -1,8 +1,8 @@
 import z from "zod";
-import paginationData from "../utils/paginationData";
-import { DEFAULT_LIMIT } from "./structure/constants";
-import MCPUserError from "./structure/MCPUserError";
-import registerTool from "./structure/registerTool";
+import paginationData from "../../utils/paginationData";
+import { DEFAULT_LIMIT } from "../structure/constants";
+import MCPUserError from "../structure/MCPUserError";
+import registerTool from "../structure/registerTool";
 
 const description = `
 List all Folders for organizing Cost Reports. Folders can be nested within other Folders via the parent_folder_token field.
