+++ src/tools/index.ts
@@ -8,17 +8,15 @@ import "./canvases";
 import "./cost-alerts";
 import "./cost-reports";
 import "./costs";
-import "./create-folder";
 import "./create-report-notification";
 import "./create-virtual-tag-config";
 import "./current-user";
 import "./dashboards";
-import "./delete-folder";
 import "./delete-report-notification";
 import "./financial-commitment-reports";
+import "./folders";
 import "./get-cost-provider-accounts";
 import "./get-financial-commitment-report";
-import "./get-folder";
 import "./get-provider-resource";
 import "./get-report-notification";
 import "./get-team";
@@ -27,7 +25,6 @@ import "./list-audit-logs";
 import "./list-cost-integrations";
 import "./list-cost-providers";
 import "./list-cost-services";
-import "./list-folders";
 import "./list-provider-resources";
 import "./list-report-notifications";
 import "./network-flow-reports";
@@ -36,7 +33,6 @@ import "./recommendations";
 import "./resource-reports";
 import "./submit-user-feedback";
 import "./tags";
-import "./update-folder";
 import "./update-report-notification";
 import "./users";
 import "./workspaces";
