+++ test/tools/folders/delete-folder.test.ts
@@ -1,17 +1,7 @@
-import { type GetFolderResponse, pathEncode } from "@vantage-sh/vantage-client";
+import { pathEncode } from "@vantage-sh/vantage-client";
 import { expect } from "vitest";
-import tool from "../../src/tools/get-folder";
-import { requestsInOrder, testTool } from "../../src/utils/testing";
-
-const success: GetFolderResponse = {
-  token: "fldr_123",
-  title: "Platform Team Reports",
-  type: "cost_reports",
-  saved_filter_tokens: [],
-  created_at: "2023-01-01T00:00:00Z",
-  updated_at: "2023-01-01T00:00:00Z",
-  workspace_token: "wrkspc_123",
-};
+import tool from "../../../src/tools/folders/delete-folder";
+import { requestsInOrder, testTool } from "../../../src/utils/testing";
 
 testTool(
   tool,
@@ -30,16 +20,16 @@ testTool(
         {
           endpoint: `/v2/folders/${pathEncode("fldr_123")}`,
           params: {},
-          method: "GET",
+          method: "DELETE",
           result: {
             ok: true,
-            data: success,
+            data: undefined,
           },
         },
       ]),
       handler: async ({ callExpectingSuccess }) => {
         const res = await callExpectingSuccess({ folder_token: "fldr_123" });
-        expect(res).toEqual(success);
+        expect(res).toEqual({ token: "fldr_123" });
       },
     },
     {
@@ -48,7 +38,7 @@ testTool(
         {
           endpoint: `/v2/folders/${pathEncode("fldr_notfound")}`,
           params: {},
-          method: "GET",
+          method: "DELETE",
           result: {
             ok: false,
             errors: [{ message: "Folder not found" }],
